requirements.txt
annotated-types==0.7.0
anyio==4.9.0
bidict==0.23.1
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
fastapi==0.115.12
Flask==3.1.1
Flask-SocketIO==5.5.1
grpcio==1.72.1
grpcio-tools==1.72.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
protobuf==6.31.1
pydantic==2.11.5
pydantic_core==2.33.2
python-engineio==4.12.1
python-multipart==0.0.20
python-socketio==5.13.0
requests==2.32.3
setuptools==80.9.0
simple-websocket==1.1.0
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.4.0
uvicorn==0.34.3
websockets==15.0.1
Werkzeug==3.1.3
wsproto==1.2.0
